requirements.txt
# 邮件导出工具依赖文件
# Python >= 3.6

# 必需依赖
# OAuth2认证支持（Gmail等）
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0

# 可选依赖
# SOCKS代理支持
PySocks>=1.7.1

# 日期选择器组件（增强GUI体验）
tkcalendar>=1.6.1

# 注意事项:
# 1. 本项目主要使用Python标准库
# 2. tkinter通常随Python安装，如果缺失请安装:
#    - Ubuntu/Debian: sudo apt-get install python3-tk
#    - CentOS/RHEL: sudo yum install tkinter
# 3. 核心功能不依赖第三方包，OAuth功能需要google-auth系列包